Palma_AI/pyproject.toml
[project]
name = "palma-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"langchain",
	"langgraph",
	"python-dotenv",
	"langchain-tavily",
	"pinecone",
	"langchain-pinecone>=0.2.13",
	"langchain-core>=1.2.11",
	"langchain-text-splitters>=1.1.0",
	"fastapi",
	"uvicorn[standard]",
	"langchain-community>=0.4.1",
	"llama-cpp-python>=0.3.16",
	"langchain-ollama>=1.0.1",
]

[dependency-groups]
dev = [
	"black",
	"isort",
]
